TransactionsServices-main/currency_converter/requirements.txt
redis
kafka-python
apscheduler
httpx